refactor(ontology): Clean up code
